IA-3433 Refac: Move tasks api to views.py and serializers.py
