+++ iaso/api/tasks/create/export_mobile_setup.py
@@ -5,7 +5,7 @@ from django.contrib.auth.password_validation import validate_password
 
 from hat.menupermissions import models as permission
 from iaso.api.common import HasPermission
-from iaso.api.tasks import TaskSerializer
+from iaso.api.tasks.serializers import TaskSerializer
 from iaso.tasks.export_mobile_app_setup_for_user import export_mobile_app_setup_for_user
